% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{cppredict}
\alias{cppredict}
\title{cppredict Function}
\usage{
cppredict(infile = "data/next_campaign_data.csv")
}
\arguments{
\item{infile}{Defaults to "next_campaign_data.csv" under the data folder}
}
\value{
a dataframe of campaign result predictions
}
\description{
This function allows you to update the "updatedbankingdata.csv" from the infile of your choice.  The infile should be ; separated csv files. The dataframe will be returned.
}
\examples{
cppredict(infile)
}
\keyword{cppredict}
