% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{cppredict}
\alias{cppredict}
\title{cppredict Function}
\usage{
cppredict(infile = "bankingclientsprofile/data/next_campaign_data.csv")
}
\arguments{
\item{infile}{Defaults to "next_campaign_data.csv" under the data folder}
}
\value{
a dataframe of campaign result predictions
}
\description{
This function will predict the result of your next campaign call.  The parameter infile should be a , separated csv files. A dataframe with the prediction will be returned.
}
\examples{
cppredict(infile)
}
\keyword{cppredict}
